MAT-8749 change endpoint
